[UPDATE]OrderItem.java의 price에 대한 타입 변경 int -> long
OrderItemCriteria.java의 getPeriod(int day)에서 day가 0일 때 오류를 수정
OrderList.java의 price 단위 알아보기 쉽게 String fomatting함
UserDao의 sns -> sms로 변수명을 변경
[ADD]modifyForm.jsp 회원정보 조회및 변경하는 페이지 만듬(미완)

+++ src/main/java/semi/vo/OrderItem.java
@@ -2,7 +2,7 @@ package semi.vo;
 
 public class OrderItem {
 	private Order order;
-	private int orderProductPrice;
+	private long orderProductPrice;
 	private int orderProductQuantity;
 	private ProductItem productItem;
 	
@@ -16,11 +16,11 @@ public class OrderItem {
 		this.order = order;
 	}
 
-	public int getOrderProductPrice() {
+	public long getOrderProductPrice() {
 		return orderProductPrice;
 	}
 
-	public void setOrderProductPrice(int orderProductPrice) {
+	public void setOrderProductPrice(long orderProductPrice) {
 		this.orderProductPrice = orderProductPrice;
 	}
